#1 : Removed create-react-app boilerplate and converted the project to typescript
