+++ src/App.tsx
@@ -1,0 +1,11 @@
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <h2>Hello World</h2> 
+    </div>
+  );
+}
+
+export default App;
